Install Redux and move components in SignUp into individual files
