+++ decadegraphy/static/js/Campaign/index.js
@@ -1,4 +1,4 @@
-import SignUp from './components/SignUp.js'
+import SignUp from './components/SignUp/index.js'
 import Message from './components/Message.js'
 
 export {SignUp, Message}
