+++ decadegraphy/static/js/Campaign/reducer.js
@@ -1,0 +1,9 @@
+const initialState = {
+
+}
+
+function reducer(previousState = initialState, { type, payload }) {
+  return previousState
+}
+
+export default reducer
