+++ decadegraphy/static/js/index.js
@@ -3,6 +3,8 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import createBrowserHistory from 'history/createBrowserHistory'
 import * as Campaign from './Campaign'
+import { Provider } from 'react-redux'
+import { store } from './store'
 
 import '../css/main.scss'
 
@@ -17,11 +19,13 @@ const history = createBrowserHistory()
 
 if (document.getElementById('app')) {
   ReactDOM.render((
-    <Router>
-      <div>
-        <Route exact path="/campaigns/signup(/)?(\d+)?" component={Campaign.SignUp} />
-        <Route exact path="/campaigns/signup/:message" component={Campaign.Message} />
-      </div>
-    </Router>
+    <Provider store={store}>
+      <Router>
+        <div>
+          <Route exact path="/campaigns/signup(/)?(\d+)?" component={Campaign.SignUp} />
+          <Route exact path="/campaigns/signup/:message" component={Campaign.Message} />
+        </div>
+      </Router>
+    </Provider>
   ), document.getElementById('app'))
 }
