+++ decadegraphy/static/js/reducer.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux'
+import campaignsReducer from './Campaign/reducer'
+
+const reducers = combineReducers({
+  campaignsReducer
+})
+
+export default reducers
